wl as int

+++ pyjeti/jeti.py
@@ -157,7 +157,9 @@ class Spectrometer:
             start_wavelength = 380
             end_wavelength = 780
             num_values = int((end_wavelength - start_wavelength) / step_nm) + 1
-            wl = np.arange(start_wavelength, end_wavelength + step_nm, step_nm)
+            wl = np.arange(start_wavelength, end_wavelength + step_nm, step_nm).astype(
+                int
+            )
             spectrum = np.random.rand(num_values).astype(np.float32) / 1000
             spd = np.vstack((wl, spectrum))
             return spd
